+++ src/Document/Content.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Symfony\Cmf\Component\Testing\Document;
 
 use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCRODM;
@@ -8,6 +18,9 @@ use Doctrine\ODM\PHPCR\Mapping\Annotations as PHPCRODM;
  * Test content document
  *
  * Very simple, referenceable document.
+ *
+ * @deprecated This Document is deprecated as of 1.1 and will be removed in 
+ * 2.0. Move the fixture to your own bundle instead.
  *
  * @PHPCRODM\Document(referenceable=true)
  */