+++ src/Functional/DbManager/PHPCR.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Symfony\Cmf\Component\Testing\Functional\DbManager;
 
 use Doctrine\Common\DataFixtures\Purger\PHPCRPurger;
@@ -25,10 +35,14 @@ class PHPCR
         return $this->container->get('doctrine_phpcr');
     }
 
-    public function getOm()
+    /**
+     * @param null|string $managerName
+     * @return DocumentManager
+     */
+    public function getOm($managerName = null)
     {
         if (!$this->om) {
-            $this->om = $this->getRegistry()->getManager();
+            $this->om = $this->getRegistry()->getManager($managerName);
         }
 
         return $this->om;