+++ src/HttpKernel/TestKernel.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Symfony\Cmf\Component\Testing\HttpKernel;
 
 use Symfony\Component\Filesystem\Filesystem;
@@ -27,28 +37,49 @@ abstract class TestKernel extends Kernel
     public function init()
     {
         $this->registerBundleSet('default', array(
-            '\Symfony\Bundle\FrameworkBundle\FrameworkBundle',
-            '\Symfony\Bundle\SecurityBundle\SecurityBundle',
-            '\Symfony\Bundle\TwigBundle\TwigBundle',
-            '\Symfony\Bundle\MonologBundle\MonologBundle',
+            'Symfony\Bundle\FrameworkBundle\FrameworkBundle',
+            'Symfony\Bundle\SecurityBundle\SecurityBundle',
+            'Symfony\Bundle\TwigBundle\TwigBundle',
+            'Symfony\Bundle\MonologBundle\MonologBundle',
         ));
 
         $this->registerBundleSet('phpcr_odm', array(
-            '\Doctrine\Bundle\DoctrineBundle\DoctrineBundle',
-            '\Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle',
+            'Doctrine\Bundle\DoctrineBundle\DoctrineBundle',
+            'Doctrine\Bundle\PHPCRBundle\DoctrinePHPCRBundle',
         ));
 
         $this->registerBundleSet('doctrine_orm', array(
-            '\Doctrine\Bundle\DoctrineBundle\DoctrineBundle',
+            'Doctrine\Bundle\DoctrineBundle\DoctrineBundle',
         ));
 
         $this->registerBundleSet('sonata_admin', array(
-            '\Sonata\BlockBundle\SonataBlockBundle',
-            '\Sonata\AdminBundle\SonataAdminBundle',
-            '\Sonata\CoreBundle\SonataCoreBundle',
-            '\Sonata\jQueryBundle\SonatajQueryBundle',
+            'Sonata\BlockBundle\SonataBlockBundle',
+            'Sonata\CoreBundle\SonataCoreBundle',
+            'Sonata\AdminBundle\SonataAdminBundle',
+            'Sonata\jQueryBundle\SonatajQueryBundle',
             'Knp\Bundle\MenuBundle\KnpMenuBundle',
-            '\Sonata\DoctrinePHPCRAdminBundle\SonataDoctrinePHPCRAdminBundle',
+            'Sonata\DoctrinePHPCRAdminBundle\SonataDoctrinePHPCRAdminBundle',
+            'Symfony\Cmf\Bundle\TreeBrowserBundle\CmfTreeBrowserBundle',
+            'FOS\JsRoutingBundle\FOSJsRoutingBundle',
+        ));
+
+        $this->registerBundleSet('sonata_admin_orm', array(
+            'Sonata\BlockBundle\SonataBlockBundle',
+            'Sonata\CoreBundle\SonataCoreBundle',
+            'Sonata\AdminBundle\SonataAdminBundle',
+            'Sonata\jQueryBundle\SonatajQueryBundle',
+            'Knp\Bundle\MenuBundle\KnpMenuBundle',
+            'FOS\JsRoutingBundle\FOSJsRoutingBundle',
+            'Sonata\DoctrineORMAdminBundle\SonataDoctrineORMAdminBundle',
+        ));
+
+        $this->registerBundleSet('sonata_admin_phpcr', array(
+            'Sonata\BlockBundle\SonataBlockBundle',
+            'Sonata\CoreBundle\SonataCoreBundle',
+            'Sonata\AdminBundle\SonataAdminBundle',
+            'Sonata\jQueryBundle\SonatajQueryBundle',
+            'Knp\Bundle\MenuBundle\KnpMenuBundle',
+            'Sonata\DoctrinePHPCRAdminBundle\SonataDoctrinePHPCRAdminBundle',
             'Symfony\Cmf\Bundle\TreeBrowserBundle\CmfTreeBrowserBundle',
             'FOS\JsRoutingBundle\FOSJsRoutingBundle',
         ));
@@ -114,7 +145,7 @@ abstract class TestKernel extends Kernel
                 ));
             }
 
-            $this->requiredBundles[] = new $bundle;
+            $this->requiredBundles[$bundle] = new $bundle;
         }
     }