+++ src/Phpunit/DatabaseTestListener.php
@@ -1,9 +1,21 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Symfony\Cmf\Component\Testing\Phpunit;
 
 use Symfony\Component\Process\ProcessBuilder;
 use Symfony\Component\Process\Process;
+use Symfony\Component\Process\ProcessUtils;
+use Symfony\Component\Process\PhpExecutableFinder;
 use Symfony\Component\HttpKernel\Kernel;
 use Doctrine\Common\DataFixtures\Purger;
 
@@ -11,12 +23,20 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
 {
     protected static $currentSuite;
     private $processBuilder;
+    private $prefix = array();
 
     public function __construct($processBuilder = null)
     {
         if (null === $processBuilder) {
-            $this->processBuilder = new PrefixedProcessBuilder();
-            $this->processBuilder->setPrefix(array('php', __DIR__.'/../../../../../../bin/console'));
+            $this->processBuilder = new ProcessBuilder();
+            $phpExecutableFinder = new PhpExecutableFinder();
+            $phpExecutable = $phpExecutableFinder->find();
+            if (false === $phpExecutable) {
+                throw new \RuntimeException('No PHP executable found on the current system.');
+            }
+
+            // Symfony 2.3 does not support array prefix, so we have to implement it ourselves
+            $this->prefix = array($phpExecutable, __DIR__.'/../../../../../../bin/console');
         } else {
             $this->processBuilder = $processBuilder;
         }
@@ -35,6 +55,10 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
     }
  
     public function addSkippedTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
+    {
+    }
+    
+    public function addRiskyTest(PHPUnit_Framework_Test $test, Exception $e, $time)
     {
     }
  
@@ -76,7 +100,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
                 break;
 
             default:
-                if (!class_exists($suite->getName())) {
+                if (!class_exists($suite->getName()) && false === strpos($suite->getName(), '::')) {
                     echo PHP_EOL.PHP_EOL.'['.$suite->getName().']'.PHP_EOL;
                 }
         }
@@ -87,7 +111,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         echo PHP_EOL.PHP_EOL;
 
         $process = $this->processBuilder
-            ->setArguments(array('doctrine:phpcr:init:dbal', '--drop'))
+            ->setArguments(array_merge($this->prefix, array('doctrine:phpcr:init:dbal', '--drop')))
             ->getProcess();
         $process->run();
 
@@ -98,7 +122,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
                     $suite->markTestSuiteSkipped('[PHPCR] Error when initializing dbal: '.$output);
                 } else {
                     $process = $this->processBuilder
-                        ->setArguments(array('doctrine:phpcr:repository:init'))
+                        ->setArguments(array_merge($this->prefix, array('doctrine:phpcr:repository:init')))
                         ->getProcess();
                     $process->run();
 
@@ -126,7 +150,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         echo PHP_EOL.PHP_EOL;
 
         $process = $this->processBuilder
-            ->setArguments(array('doctrine:schema:drop', '--env=orm', '--force'))
+            ->setArguments(array_merge($this->prefix, array('doctrine:schema:drop', '--env=orm', '--force')))
             ->getProcess();
         $process->run();
 
@@ -142,7 +166,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         }
 
         $process = $this->processBuilder
-            ->setArguments(array('doctrine:database:create', '--env=orm'))
+            ->setArguments(array_merge($this->prefix, array('doctrine:database:create', '--env=orm')))
             ->getProcess();
         $process->run();
 
@@ -153,7 +177,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
                     $suite->markTestSuiteSkipped('[ORM] Error when creating database: '.$output);
                 } else {
                     $process = $this->processBuilder
-                        ->setArguments(array('doctrine:schema:create', '--env=orm'))
+                        ->setArguments(array_merge($this->prefix, array('doctrine:schema:create', '--env=orm')))
                         ->getProcess();
                     $process->run();
 
@@ -183,7 +207,7 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
         }
 
         $process = $this->processBuilder
-            ->setArguments(array('doctrine:database:drop', '--force'))
+            ->setArguments(array_merge($this->prefix, array('doctrine:database:drop', '--force')))
             ->getProcess();
         $process->run();
 
@@ -196,54 +220,5 @@ class DatabaseTestListener implements \PHPUnit_Framework_TestListener
             }
             break;
         }
-    }
-}
-
-class PrefixedProcessBuilder extends ProcessBuilder
-{
-    private $prefixes = array();
-    private $activated = true;
-
-    public function __construct(array $arguments = array())
-    {
-        parent::__construct($arguments);
-
-        if (class_exists('Symfony\Component\HttpKernel\Kernel')) {
-            $this->activated = Kernel::VERSION_ID <= 20400;
-        } else {
-            $ref = new \ReflectionMethod(get_parent_class($this), 'setPrefix'); 
-            $doc = $ref->getDocComment();
-
-            $this->activated = false === strpos('array', $doc);
-        }
-    }
-
-    public function setPrefix($prefix)
-    {
-        $prefixes = is_array($prefix) ? $prefix : array($prefix);
-
-        if (!$this->activated) {
-            return parent::setPrefix($prefix);
-        }
-
-        foreach ($prefixes as $prefix) {
-            $this->prefixes[] = $prefix;
-        }
-        
-        return $this;
-    }
-
-    public function getProcess()
-    {
-        if (!$this->activated || 0 === count($this->prefixes)) {
-            return parent::getProcess();
-        }
-
-        $process = parent::getProcess();
-        $command = implode(' ', array_map('escapeshellarg', $this->prefixes)).' '.$process->getCommandLine();
-
-        $process->setCommandLine($command);
-
-        return $process;
     }
 }