+++ src/Unit/XmlSchemaTestCase.php
@@ -1,8 +1,18 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Symfony\Cmf\Component\Testing\Unit;
 
-class XmlSchemaTestCase extends \PHPUnit_Framework_TestCase
+abstract class XmlSchemaTestCase extends \PHPUnit_Framework_TestCase
 {
     public static function assertSchemaAcceptsXml($xmlDoms, $schemaPath, $message = '')
     {