+++ tests/HttpKernel/TestKernelTest.php
@@ -1,5 +1,15 @@
 <?php
 
+/*
+ * This file is part of the Symfony CMF package.
+ *
+ * (c) 2011-2014 Symfony CMF
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+
 namespace Tests\HttpKernel;
 
 use Symfony\Cmf\Component\Testing\HttpKernel\TestKernel;
@@ -17,14 +27,24 @@ class TestKernelTest extends \PHPUnit_Framework_TestCase
         );
     }
 
-    public function testBundleSetRequire()
+    /**
+     * @dataProvider bundleSetProvider
+     */
+    public function testBundleSetRequire(array $bundleSets, $count)
     {
         $this->kernel->init();
-        $this->kernel->requireBundleSets(array(
-            'default', 'phpcr_odm'
-        ));
+        $this->kernel->requireBundleSets($bundleSets);
         $bundles = $this->kernel->registerBundles();
-        $this->assertCount(6, $bundles);
+        $this->assertCount($count, $bundles);
+    }
+
+    public function bundleSetProvider()
+    {
+        return array(
+            array(array('default', 'phpcr_odm'), 6),
+            array(array('default', 'doctrine_orm'), 5),
+            array(array('default', 'doctrine_orm', 'phpcr_odm'), 6),
+        );
     }
 
     public function testBundleAdd()