+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 Changelog
 =========
 
-* **2013-12-27**: Added XmlSchemaTestCase to test XML schema's
-* **2013-11-17**: Added DatabaseTestListener to support database testing
+* **2014-06-06**: Updated to PSR-4 autoloading
+
+1.1.0-RC2
+---------
+
+* **2014-06-11**: fetch manager by its name
+* **2014-04-24**: [DEPRECATE] Deprecated the `sonata_admin` bundle set
+* **2014-04-24**: added bundle sets: `sonata_admin_orm` and `sonata_admin_phpcr`
+* **2014-04-11**: dropped Symfony 2.2 compatibility
+
+1.1.0-RC1
+---------
+
+* **2013-04-07**: [DEPRECATE] Deprecated `LoadBaseData` DataFixture, it will be removed in 2.0
+* **2013-04-02**: [DEPRECATE] Deprecated `Content` document, it will be removed in 2.0
+* **2013-12-27**: Added `XmlSchemaTestCase` to test XML schema's
+* **2013-11-17**: Added `DatabaseTestListener` to support database testing